TDD_Project/lib/diary_class_entry.py
# File: lib/diary_class_entry.py

class DiaryEntry:
    # Public Properties:
    #   title: a string
    #   contents: a string

    def __init__(self, title, contents): # title, contents are strings
        # Side-effects:
        #   Sets the title and contents properties
        self.title = title
        self.contents = contents
        self.chunk_position = 0


    def count_words(self):
        # Returns:
        #   An integer representing the number of words in the contents
        return len(self.contents.split())

    def reading_time(self, wpm):
        # Parameters:
        #   wpm: an integer representing the number of words the user can read
        #        per minute
        # Returns:
        #   An integer representing an estimate of the reading time in minutes
        #   for the contents at the given wpm.
        word_total = self.count_words()
        return (word_total / wpm)

    def reading_chunk(self, wpm, minutes):
        # Parameters:
        #   wpm: an integer representing the number of words the user can read
        #        per minute
        #   minutes: an integer representing the number of minutes the user has
        #            to read
        # Returns:
        #   A string representing a chunk of the contents that the user could
        #   read in the given number of minutes.
        # If called again, `reading_chunk` should return the next chunk,
        # skipping what has already been read, until the contents is fully read.
        # The next call after that it should restart from the beginning.
        wpm_multiplied_minutes = wpm * minutes
        contents_list = self.contents.split()

        current = self.chunk_position
        future = current + wpm_multiplied_minutes

        if self.chunk_position >= len(self.contents):
            self.chunk_position = 0
        
        next_chunk = (" ".join(contents_list[current:future]))
        self.chunk_position = future
        return next_chunk
